1. Fix address-sanitize problems, 2. a workaroud for sendrecv comm type, 3. a workaround for set dim = 1.

+++ astra-sim/system/CSVWriter.cc
@@ -188,8 +188,8 @@ void CSVWriter::write_cell(int row, int column, std::string data) {
 
   do {
     lseek(fildes, 0, SEEK_SET);
-    status = write(fildes, str.c_str(), str.length());
-  } while (status != str.length());
+    status = write(fildes, str.c_str(), static_cast<int>(str.length()));
+  } while (status != static_cast<int>(str.length()));
 
   do {
     status = lseek(fildes, 0, SEEK_SET);
